create autocomplete component, make a delay to be set by props
