+++ src/components/Autocomplete/index.ts
@@ -1,0 +1,1 @@
+export * from './Autocomplete';
